重生之我复活了
我复活了，复活在了项目截止的前5天，这一世，我要拿回我的一切

+++ SonicStrike/Assets/Scripts/itemSystem/shield.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class shield
 {
     public shieldData basicData;
-    public float durability;
+    public int durability;
 }
